+++ libraries/GAMMA/gamma.cpp
@@ -1,10 +1,10 @@
 //
 //    FILE: gamma.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.4.1
+// VERSION: 0.5.0
 //    DATE: 2020-08-08
 // PURPOSE: Arduino Library to efficiently hold a gamma lookup table
-
+//     URL: https://github.com/RobTillaart/GAMMA
 
 #include "gamma.h"
 
@@ -24,13 +24,13 @@ GAMMA::GAMMA(uint16_t size)
   }
   _mask = (1 << _shift) - 1;
   _interval = GAMMA_MAX_SIZE / _size;
-};
+}
 
 
 GAMMA::~GAMMA()
 {
   if (_table) free(_table);
-};
+}
 
 
 bool GAMMA::begin()
@@ -39,19 +39,36 @@ bool GAMMA::begin()
   {
     _table = (uint8_t *) malloc(_size + 1);
   }
-  if (_table == NULL) return false;
+  if (_table == NULL)
+  {
+    _size = 0;
+    return false;
+  }
   setGamma(2.8);
   return true;
-};
+}
+
+
+bool GAMMA::isAllocated()
+{
+  return (_table != NULL);
+}
 
 
 bool GAMMA::setGamma(float gamma)
 {
+  _negative = false;
   if (_table == NULL) return false;
-  if (gamma <= 0) return false;
+  if (gamma == 0) return false;
+  if (gamma < 0)
+  {
+    _negative = true;
+    gamma = -gamma;
+  }
   if (_gamma != gamma)
   {
-    yield();  //  try to keep ESP happy
+    //  try to keep ESP happy
+    yield();
     _gamma = gamma;
 #if defined(ESP32)
     //  confirmed faster for ESP32.
@@ -70,17 +87,19 @@ bool GAMMA::setGamma(float gamma)
       _table[i] = pow(i * tmp, _gamma) * 255 + 0.444;
     }
 #endif
+    //  handle experimental negative values.
     _table[0] = 0;
     _table[_size] = 255;  //  anchor for interpolation.
   }
   return true;
-};
+}
 
 
 float GAMMA::getGamma()
 {
+  if (_negative) return -_gamma;
   return _gamma;
-};
+}
 
 
 uint8_t GAMMA::operator[] (uint8_t index)
@@ -94,20 +113,29 @@ uint8_t GAMMA::operator[] (uint8_t index)
   uint32_t  i = index >> _shift;
   uint32_t  m = index & _mask;
   //  exact element shortcut
-  if ( m == 0 ) return _table[i];
+  if ( m == 0 )
+  {
+    if (_negative) return 255 - _table[i];
+    return _table[i];
+  }
 
   //  interpolation
   //  delta must be uint16_t to prevent overflow. (small tables)
   //        delta * m can be > 8 bit.
   uint32_t delta = _table[i+1] - _table[i];
   delta = ( delta * m + _interval/2 ) >> _shift;  //  == /_interval;
+  if (_negative) return 255 - (_table[i] + delta);
   return _table[i] + delta;
 
 #else
   uint16_t  i = index >> _shift;
   uint16_t  m = index & _mask;
   //  exact element shortcut
-  if ( m == 0 ) return _table[i];
+  if ( m == 0 )
+  {
+    if (_negative) return 255 - _table[i];
+    return _table[i];
+  }
 
   //  interpolation
   //  delta must be uint16_t to prevent overflow. (small tables)
@@ -117,19 +145,20 @@ uint8_t GAMMA::operator[] (uint8_t index)
   if (delta != 0)
   {
     // delta = ( delta * m + _interval/2 ) >> _shift;  //  == /_interval;
-    delta += m;
-    delta +=_interval/2;
+    delta *= m;
+    delta +=_interval /2;
     delta >>= _shift;  //  == /_interval;
   }
+  if (_negative) return 255 - (_table[i] + delta);
   return _table[i] + delta;
 #endif
-};
+}
 
 
 uint16_t GAMMA::size()
 {
   return _size + 1;
-};
+}
 
 
 uint16_t GAMMA::distinct()
@@ -143,7 +172,7 @@ uint16_t GAMMA::distinct()
     count++;
   }
   return count;
-};
+}
 
 
 bool GAMMA::dump(Stream *str)
@@ -154,7 +183,7 @@ bool GAMMA::dump(Stream *str)
     str->println(_table[i]);
   }
   return true;
-};
+}
 
 
 bool GAMMA::dumpArray(Stream *str)
@@ -173,7 +202,7 @@ bool GAMMA::dumpArray(Stream *str)
   }
   str->print("\n  };\n\n");
   return true;
-};
+}
 
 
 //  performance investigation