+++ libraries/GAMMA/gamma.h
@@ -2,14 +2,15 @@
 //
 //    FILE: gamma.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.4.1
+// VERSION: 0.5.0
 //    DATE: 2020-08-08
 // PURPOSE: Arduino Library to efficiently hold a gamma lookup table
+//     URL: https://github.com/RobTillaart/GAMMA
 
 
 #include "Arduino.h"
 
-#define GAMMA_LIB_VERSION                 (F("0.4.1"))
+#define GAMMA_LIB_VERSION                 (F("0.5.0"))
 
 #define GAMMA_DEFAULT_SIZE                32
 #define GAMMA_MAX_SIZE                    256
@@ -27,9 +28,11 @@ public:
   //  sets default gamma = 2.8
   //  Returns false if allocation fails
   bool begin();
+  bool isAllocated();
 
   //  CORE
-  //  Returns false if gamma <= 0
+  //  Returns false if gamma == 0
+  //  (experimental for gamma < 0)
   bool  setGamma(float gamma);
   float getGamma();
   //  access values with index operator
@@ -52,6 +55,7 @@ private:
   uint8_t   _interval = 0;
   float     _gamma    = 1.0;  //  1.0 == no gamma, linear.
   uint8_t * _table    = NULL;
+  bool      _negative = false;
 
   float     fastPow(float a, float b);
 };